// ts/tests/support/echo_pb.ts
// @generated by protoc-gen-es v2.13.0 with parameter "target=ts"
// @generated from file echo.proto (package echo, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file echo.proto.
 */
export const file_echo: GenFile = /*@__PURE__*/
  fileDesc("CgplY2hvLnByb3RvEgRlY2hvIhQKBEVjaG8SDAoEdGV4dBgBIAEoCTKcAQoLRWNob1NlcnZpY2USHwoFVW5hcnkSCi5lY2hvLkVjaG8aCi5lY2hvLkVjaG8SHgoERmFpbBIKLmVjaG8uRWNobxoKLmVjaG8uRWNobxIoCgxTZXJ2ZXJTdHJlYW0SCi5lY2hvLkVjaG8aCi5lY2hvLkVjaG8wARIiCgRCaWRpEgouZWNoby5FY2hvGgouZWNoby5FY2hvKAEwAWIGcHJvdG8z");

/**
 * @generated from message echo.Echo
 */
export type Echo = Message<"echo.Echo"> & {
  /**
   * @generated from field: string text = 1;
   */
  text: string;
};

/**
 * Describes the message echo.Echo.
 * Use `create(EchoSchema)` to create a new message.
 */
export const EchoSchema: GenMessage<Echo> = /*@__PURE__*/
  messageDesc(file_echo, 0);

/**
 * @generated from service echo.EchoService
 */
export const EchoService: GenService<{
  /**
   * @generated from rpc echo.EchoService.Unary
   */
  unary: {
    methodKind: "unary";
    input: typeof EchoSchema;
    output: typeof EchoSchema;
  },
  /**
   * @generated from rpc echo.EchoService.Fail
   */
  fail: {
    methodKind: "unary";
    input: typeof EchoSchema;
    output: typeof EchoSchema;
  },
  /**
   * @generated from rpc echo.EchoService.ServerStream
   */
  serverStream: {
    methodKind: "server_streaming";
    input: typeof EchoSchema;
    output: typeof EchoSchema;
  },
  /**
   * @generated from rpc echo.EchoService.Bidi
   */
  bidi: {
    methodKind: "bidi_streaming";
    input: typeof EchoSchema;
    output: typeof EchoSchema;
  },
}> = /*@__PURE__*/
  serviceDesc(file_echo, 0);